refactor(Service): 문서 교체로 인한 Service 상호간 함수 교체
